refactor: entity
