+++ src/main/java/id/ac/ui/cs/advprog/hiringgo/entity/Log.java
@@ -9,8 +9,8 @@ import lombok.Data;
 import java.time.LocalDate;
 import java.time.LocalTime;
 
-import id.ac.ui.cs.advprog.hiringgo.manajemenLog.model.enums.StatusLog;
-import id.ac.ui.cs.advprog.hiringgo.manajemenLog.model.enums.TipeKategori;
+import id.ac.ui.cs.advprog.hiringgo.manajemenLog.enums.StatusLog;
+import id.ac.ui.cs.advprog.hiringgo.manajemenLog.enums.TipeKategori;
 
 @Entity
 @Table(name = "log")
