+++ src/main/java/id/ac/ui/cs/advprog/hiringgo/entity/PendaftarLowongan.java
@@ -2,11 +2,7 @@ package id.ac.ui.cs.advprog.hiringgo.entity;
 
 import java.util.UUID;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -19,16 +15,17 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
+@Table(name = "mahasiswa_pendaftar_lowongan")
 public class PendaftarLowongan {
     @Id
-    @GeneratedValue
-    private UUID id;
+    private String id;
 
     @ManyToOne
     @JoinColumn(name = "lowongan_id", nullable = false)
-    private id.ac.ui.cs.advprog.hiringgo.entity.Lowongan lowongan;
+    private Lowongan lowongan;
 
     @ManyToOne
+    @JoinColumn(name = "mahasiswa_id", nullable = false)
     private Mahasiswa mahasiswa;
 
     private int jumlahSks;
