+++ src/main/java/id/ac/ui/cs/advprog/hiringgo/manajemenLog/dto/LogRequest.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 import java.time.LocalDate;
 import java.time.LocalTime;
 
-import id.ac.ui.cs.advprog.hiringgo.manajemenLog.model.enums.TipeKategori;
+import id.ac.ui.cs.advprog.hiringgo.manajemenLog.enums.TipeKategori;
 
 @Data
 @Builder
