+++ src/main/java/id/ac/ui/cs/advprog/hiringgo/manajemenLog/service/LogServiceImpl.java
@@ -3,8 +3,9 @@ package id.ac.ui.cs.advprog.hiringgo.manajemenLog.service;
 import id.ac.ui.cs.advprog.hiringgo.entity.Lowongan;
 import id.ac.ui.cs.advprog.hiringgo.entity.MataKuliah;
 import id.ac.ui.cs.advprog.hiringgo.entity.Mahasiswa;
-import id.ac.ui.cs.advprog.hiringgo.manajemenLog.repository.AsdosMataKuliahRepository;
+import id.ac.ui.cs.advprog.hiringgo.repository.AsdosMataKuliahRepository;
 import id.ac.ui.cs.advprog.hiringgo.manajemenlowongan.repository.LowonganRepository;
+import id.ac.ui.cs.advprog.hiringgo.repository.LogRepository;
 import id.ac.ui.cs.advprog.hiringgo.repository.MahasiswaRepository;
 import id.ac.ui.cs.advprog.hiringgo.repository.MataKuliahRepository;
 import org.springframework.stereotype.Service;
@@ -24,8 +25,7 @@ import id.ac.ui.cs.advprog.hiringgo.manajemenLog.dto.LogResponse;
 import id.ac.ui.cs.advprog.hiringgo.manajemenLog.exception.InvalidLogException;
 import id.ac.ui.cs.advprog.hiringgo.manajemenLog.exception.LogNotFoundException;
 import id.ac.ui.cs.advprog.hiringgo.entity.Log;
-import id.ac.ui.cs.advprog.hiringgo.manajemenLog.model.enums.StatusLog;
-import id.ac.ui.cs.advprog.hiringgo.manajemenLog.repository.LogRepository;
+import id.ac.ui.cs.advprog.hiringgo.manajemenLog.enums.StatusLog;
 import id.ac.ui.cs.advprog.hiringgo.manajemenLog.validation.LogValidator;
 import id.ac.ui.cs.advprog.hiringgo.manajemenLog.validation.LogValidatorFactory;
