+++ src/test/java/id/ac/ui/cs/advprog/hiringgo/manajemenLog/service/LogServiceTest.java
@@ -18,12 +18,11 @@ import id.ac.ui.cs.advprog.hiringgo.manajemenLog.dto.LogResponse;
 import id.ac.ui.cs.advprog.hiringgo.manajemenLog.exception.InvalidLogException;
 import id.ac.ui.cs.advprog.hiringgo.manajemenLog.exception.LogNotFoundException;
 import id.ac.ui.cs.advprog.hiringgo.entity.Log;
-import id.ac.ui.cs.advprog.hiringgo.manajemenLog.model.enums.StatusLog;
-import id.ac.ui.cs.advprog.hiringgo.manajemenLog.model.enums.TipeKategori;
-import id.ac.ui.cs.advprog.hiringgo.manajemenLog.repository.LogRepository;
+import id.ac.ui.cs.advprog.hiringgo.manajemenLog.enums.StatusLog;
+import id.ac.ui.cs.advprog.hiringgo.manajemenLog.enums.TipeKategori;
 import id.ac.ui.cs.advprog.hiringgo.manajemenLog.validation.LogValidator;
 import id.ac.ui.cs.advprog.hiringgo.manajemenLog.validation.LogValidatorFactory;
-import id.ac.ui.cs.advprog.hiringgo.manajemenLog.repository.AsdosMataKuliahRepository;
+import id.ac.ui.cs.advprog.hiringgo.repository.AsdosMataKuliahRepository;
 import id.ac.ui.cs.advprog.hiringgo.manajemenlowongan.repository.LowonganRepository;
 import id.ac.ui.cs.advprog.hiringgo.repository.MahasiswaRepository;
 import id.ac.ui.cs.advprog.hiringgo.repository.MataKuliahRepository;
